refactor: eliminate 40-prop waterfall in library components
- Introduce LibraryContext to hold shared library state

- PlaylistSelection/index.tsx provides context instead of threading props

- LibraryMainContent, PlaylistGrid, AlbumGrid consume from useLibraryContext()

- No change in user-visible behavior

Closes #593
